refactor breakdown ClientTest in multiple files
