+++ tests/aws-cpp-sdk-core-tests/aws/client/AWSClientTest.cpp
@@ -18,7 +18,7 @@
 #include <aws/core/platform/FileSystem.h>
 #include <aws/core/auth/AWSCredentialsProvider.h>
 #include <aws/core/platform/Environment.h>
-#include <aws/core/auth/AWSCredentialsProvider.h>
+
 #include <fstream>
 #include <thread>
 #include <aws/core/utils/logging/LogMacros.h>
@@ -179,62 +179,6 @@ TEST_F(AWSClientTestSuite, TestCreateHttpRequestWithIpV6KeepsEndpointWhenPortInU
     Aws::String parsedUri = azRequest->GetURIString();
     ASSERT_STREQ("http://[fd00:ec2::254]:8080/latest/meta-data/placement/availability-zone", parsedUri.c_str());
 }
-
-
-class AWSConfigTestSuite : public ::testing::Test
-{
-protected:
-    void SetUp()
-    {
-        SaveEnvironmentVariable("AWS_CONFIG_FILE");
-        SaveEnvironmentVariable("AWS_DEFAULT_PROFILE");
-        SaveEnvironmentVariable("AWS_PROFILE");
-        SaveEnvironmentVariable("AWS_DEFAULT_REGION");
-        SaveEnvironmentVariable("AWS_REGION");
-        SaveEnvironmentVariable("AWS_EC2_METADATA_SERVICE_ENDPOINT");
-
-        Aws::StringStream ss;
-        ss << Aws::Auth::GetConfigProfileFilename() + "_blah" << std::this_thread::get_id();
-        m_configFileName = ss.str();
-        Aws::Environment::SetEnv("AWS_CONFIG_FILE", m_configFileName.c_str(), 1);
-        Aws::Environment::UnSetEnv("AWS_DEFAULT_PROFILE");
-        Aws::Environment::UnSetEnv("AWS_PROFILE");
-        Aws::Environment::UnSetEnv("AWS_DEFAULT_REGION");
-        Aws::Environment::UnSetEnv("AWS_REGION");
-        Aws::Environment::UnSetEnv("AWS_EC2_METADATA_SERVICE_ENDPOINT");
-
-        auto profileDirectory = Aws::Auth::ProfileConfigFileAWSCredentialsProvider::GetProfileDirectory();
-        Aws::FileSystem::CreateDirectoryIfNotExists(profileDirectory.c_str());
-    }
-
-    void TearDown()
-    {
-        RestoreEnvironmentVariables();
-     }
-
-    void SaveEnvironmentVariable(const char* variableName)
-    {
-        m_environment.emplace_back(variableName, Aws::Environment::GetEnv(variableName));
-    }
-
-    void RestoreEnvironmentVariables()
-    {
-        for(const auto& iter : m_environment)
-        {
-            if(iter.second.empty())
-            {
-                Aws::Environment::UnSetEnv(iter.first);
-            }
-            else
-            {
-                Aws::Environment::SetEnv(iter.first, iter.second.c_str(), 1/*override*/);
-            }
-        }
-    }
-
-    Aws::Vector<std::pair<const char*, Aws::String>> m_environment;
-    Aws::String m_configFileName;
-};
 
 TEST_F(AWSClientTestSuite, TestValidateInvalidURI)
 {
@@ -758,53 +702,7 @@ TEST(AWSClientTest, TestOverflowContainer)
     ASSERT_STREQ(enumValue, container->RetrieveOverflow(hashcode).c_str());
 }
 
-TEST_F(AWSConfigTestSuite, TestClientConfigurationWithNonExistentProfile)
-{
-    // create a config file with profile named Dijkstra
-    Aws::OFStream configFileNew(m_configFileName.c_str(), Aws::OFStream::out | Aws::OFStream::trunc);
-    configFileNew << "[Dijkstra]" << std::endl;
-    configFileNew << "region = " << Aws::Region::US_WEST_2 << std::endl;
 
-    configFileNew.flush();
-    configFileNew.close();
-    Aws::Config::ReloadCachedConfigFile();
-
-    Aws::Client::ClientConfiguration config("Edsger");
-    EXPECT_EQ(Aws::Region::US_EAST_1, config.region);
-    EXPECT_STREQ("default", config.profileName.c_str());
-
-    // cleanup
-    Aws::FileSystem::RemoveFileIfExists(m_configFileName.c_str());
-}
-
-TEST_F(AWSConfigTestSuite, TestClientConfigurationWithNonExistentConfigFile)
-{
-    Aws::Environment::SetEnv("AWS_CONFIG_FILE", "WhatAreTheChances", 1/*overwrite*/);
-    Aws::Config::ReloadCachedConfigFile();
-
-    Aws::Client::ClientConfiguration config("default");
-    EXPECT_EQ(Aws::Region::US_EAST_1, config.region);
-    EXPECT_STREQ("default", config.profileName.c_str());
-}
-
-TEST_F(AWSConfigTestSuite, TestClientConfigurationSetsRegionToProfile)
-{
-    // create a config file with profile named Dijkstra
-    Aws::OFStream configFileNew(m_configFileName.c_str(), Aws::OFStream::out | Aws::OFStream::trunc);
-    configFileNew << "[profile Dijkstra]" << std::endl;  // profile keyword is mandatory per specification
-    configFileNew << "region = " << Aws::Region::US_WEST_2 << std::endl;
-
-    configFileNew.flush();
-    configFileNew.close();
-    Aws::Config::ReloadCachedConfigFile();
-
-    Aws::Client::ClientConfiguration config("Dijkstra");
-    EXPECT_EQ(Aws::Region::US_WEST_2, config.region);
-    EXPECT_STREQ("Dijkstra", config.profileName.c_str());
-
-    // cleanup
-    Aws::FileSystem::RemoveFileIfExists(m_configFileName.c_str());
-}
 
 class AWSRegionTest : public ::testing::Test
 {
@@ -990,155 +888,3 @@ TEST_F(AWSRegionTest, TestResolveDefaultRegion)
 }
 
 
-class AWSMetadataEndpointTestSuite : public ::testing::Test
-{
-public:
-    void SetUp()
-    {
-        SaveEnvironmentVariable("AWS_CONFIG_FILE");
-        SaveEnvironmentVariable("AWS_DEFAULT_PROFILE");
-        SaveEnvironmentVariable("AWS_PROFILE");
-        SaveEnvironmentVariable("AWS_DEFAULT_REGION");
-        SaveEnvironmentVariable("AWS_REGION");
-        SaveEnvironmentVariable("AWS_EC2_METADATA_DISABLED");
-        SaveEnvironmentVariable("AWS_EC2_METADATA_SERVICE_ENDPOINT");
-
-        Aws::StringStream ss;
-        ss << Aws::Auth::GetConfigProfileFilename() + "_blah" << std::this_thread::get_id();
-        m_configFileName = ss.str();
-        Aws::Environment::SetEnv("AWS_CONFIG_FILE", m_configFileName.c_str(), 1);
-        Aws::Environment::UnSetEnv("AWS_DEFAULT_PROFILE");
-        Aws::Environment::UnSetEnv("AWS_PROFILE");
-        Aws::Environment::UnSetEnv("AWS_DEFAULT_REGION");
-        Aws::Environment::UnSetEnv("AWS_REGION");
-        Aws::Environment::UnSetEnv("AWS_EC2_METADATA_SERVICE_ENDPOINT");
-
-        auto profileDirectory = ProfileConfigFileAWSCredentialsProvider::GetProfileDirectory();
-        Aws::FileSystem::CreateDirectoryIfNotExists(profileDirectory.c_str());
-    }
-
-    void TearDown()
-    {
-        RestoreEnvironmentVariables();
-    }
-
-    void SaveEnvironmentVariable(const char* variableName)
-    {
-        m_environment.emplace_back(variableName, Aws::Environment::GetEnv(variableName));
-    }
-
-    void RestoreEnvironmentVariables()
-    {
-        for(const auto& iter : m_environment)
-        {
-            if(iter.second.empty())
-            {
-                Aws::Environment::UnSetEnv(iter.first);
-            }
-            else
-            {
-                Aws::Environment::SetEnv(iter.first, iter.second.c_str(), 1);
-            }
-        }
-    }
-
-    Aws::Vector<std::pair<const char*, Aws::String>> m_environment;
-    Aws::String m_configFileName;
-    std::shared_ptr<MockHttpClient> mockHttpClient;
-    std::shared_ptr<MockHttpClientFactory> mockHttpClientFactory;
-};
-
-TEST_F(AWSMetadataEndpointTestSuite, TestUndeclaredEndpointForEC2MetadataInEnvUsesDefaultURI)
-{
-    Aws::Internal::CleanupEC2MetadataClient();
-
-    Aws::Environment::UnSetEnv("AWS_EC2_METADATA_SERVICE_ENDPOINT");
-    Aws::Environment::UnSetEnv("AWS_EC2_METADATA_SERVICE_ENDPOINT_MODE");
-
-    Aws::Internal::InitEC2MetadataClient();
-
-    auto client = Aws::Internal::GetEC2MetadataClient();
-    EXPECT_EQ("http://169.254.169.254", client->GetEndpoint());
-}
-
-TEST_F(AWSMetadataEndpointTestSuite, TestEndpointForEC2MetadataIsReadFromEnvWhenDeclared)
-{
-    Aws::Internal::CleanupEC2MetadataClient();
-
-    Aws::Environment::SetEnv("AWS_EC2_METADATA_SERVICE_ENDPOINT", "http://[::1]", 1/*overwrite*/);
-    Aws::Environment::UnSetEnv("AWS_EC2_METADATA_SERVICE_ENDPOINT_MODE");
-
-    Aws::Internal::InitEC2MetadataClient();
-
-    auto client = Aws::Internal::GetEC2MetadataClient();
-    EXPECT_EQ("http://[::1]", client->GetEndpoint());
-}
-
-TEST_F(AWSMetadataEndpointTestSuite, TestEndpointForEC2MetadataCanBeOverriddenBySet)
-{
-    Aws::Internal::CleanupEC2MetadataClient();
-
-    Aws::Environment::SetEnv("AWS_EC2_METADATA_SERVICE_ENDPOINT", "http://[::1]", 1/*overwrite*/);
-    Aws::Environment::UnSetEnv("AWS_EC2_METADATA_SERVICE_ENDPOINT_MODE");
-
-    Aws::Internal::InitEC2MetadataClient();
-
-    auto client = Aws::Internal::GetEC2MetadataClient();
-    EXPECT_EQ("http://[::1]", client->GetEndpoint());
-
-    client->SetEndpoint("http://127.0.0.1");
-    EXPECT_EQ("http://127.0.0.1", client->GetEndpoint());
-}
-
-TEST_F(AWSMetadataEndpointTestSuite, TestEndpointForIpV6ModeIsSetWhenNoExplicitlyOverridden)
-{
-    Aws::Internal::CleanupEC2MetadataClient();
-
-    Aws::Environment::UnSetEnv("AWS_EC2_METADATA_SERVICE_ENDPOINT");
-    Aws::Environment::SetEnv("AWS_EC2_METADATA_SERVICE_ENDPOINT_MODE", "ipV6", 1/*overwrite*/);
-
-    Aws::Internal::InitEC2MetadataClient();
-
-    auto client = Aws::Internal::GetEC2MetadataClient();
-    EXPECT_EQ("http://[fd00:ec2::254]", client->GetEndpoint());
-}
-
-TEST_F(AWSMetadataEndpointTestSuite, TestEndpointForIpV4ModeIsSetWhenNoExplicitlyOverridden)
-{
-    Aws::Internal::CleanupEC2MetadataClient();
-
-    Aws::Environment::UnSetEnv("AWS_EC2_METADATA_SERVICE_ENDPOINT");
-    Aws::Environment::SetEnv("AWS_EC2_METADATA_SERVICE_ENDPOINT_MODE", "ipV4", 1/*overwrite*/);
-
-    Aws::Internal::InitEC2MetadataClient();
-
-    auto client = Aws::Internal::GetEC2MetadataClient();
-
-    EXPECT_EQ("http://169.254.169.254", client->GetEndpoint());
-}
-
-TEST_F(AWSMetadataEndpointTestSuite, TestEndpointForIpV6ModeIsIgnoredSetWhenExplicitlyOverridden)
-{
-    Aws::Internal::CleanupEC2MetadataClient();
-
-    Aws::Environment::SetEnv("AWS_EC2_METADATA_SERVICE_ENDPOINT", "http://[::1]", 1/*overwrite*/);
-    Aws::Environment::SetEnv("AWS_EC2_METADATA_SERVICE_ENDPOINT_MODE", "ipV6", 1/*overwrite*/);
-
-    Aws::Internal::InitEC2MetadataClient();
-
-    auto client = Aws::Internal::GetEC2MetadataClient();
-    EXPECT_EQ("http://[::1]", client->GetEndpoint());
-}
-
-TEST_F(AWSMetadataEndpointTestSuite, TestEndpointForIpV4ModeIsIgnoredSetWhenExplicitlyOverridden)
-{
-    Aws::Internal::CleanupEC2MetadataClient();
-
-    Aws::Environment::SetEnv("AWS_EC2_METADATA_SERVICE_ENDPOINT", "http://[::1]", 1/*overwrite*/);
-    Aws::Environment::SetEnv("AWS_EC2_METADATA_SERVICE_ENDPOINT_MODE", "ipV4", 1/*overwrite*/);
-
-    Aws::Internal::InitEC2MetadataClient();
-
-    auto client = Aws::Internal::GetEC2MetadataClient();
-    EXPECT_EQ("http://[::1]", client->GetEndpoint());
-}
